NEXUS-7521: introduce AssetHasher API, support both direct use of AssetHashers values and injection of them as @Named implementations of AssetHasher
